+++ aiida_quantumespresso/parsers/parse_raw/neb.py
@@ -4,7 +4,7 @@ A collection of function that are used to parse the output of Quantum Espresso N
 The function that needs to be called from outside is parse_raw_output_neb().
 The functions mostly work without aiida specific functionalities.
 The parsing will try to convert whatever it can in some dictionary, which
-by operative decision doesn't have much structure encoded, [the values are simple ] 
+by operative decision doesn't have much structure encoded, [the values are simple ]
 """
 from __future__ import absolute_import
 import xml.dom.minidom
@@ -12,7 +12,7 @@ import os
 import string
 from aiida_quantumespresso.parsers import QEOutputParsingError, get_parser_info
 from aiida_quantumespresso.parsers.constants import ry_to_ev,hartree_to_ev,bohr_to_ang,ry_si,bohr_si
-from aiida_quantumespresso.parsers.raw_parser_pw import convert_qe_time_to_sec
+from aiida_quantumespresso.parsers.parse_raw.pw import convert_qe_time_to_sec
 import six
 from six.moves import range
 
@@ -20,16 +20,16 @@ def parse_raw_output_neb(out_file, input_dict,parser_opts=None):
     """
     Parses the output of a neb calculation
     Receives in input the paths to the output file.
-    
+
     :param out_file: path to neb std output
     :param input_dict: dictionary with the neb input parameters
     :param parser_opts: not used
-    
+
     :return parameter_data: a dictionary with parsed parameters
     :return iteration_data: a dictionary with arrays (for relax & md calcs.)
     :return structure_data: a dictionary with data for the output structure
     :return job_successful: a boolean that is False in case of failed calculations
-            
+
     :raises QEOutputParsingError: for errors in the parsing,
     :raises AssertionError: if two keys in the parsed dicts are found to be qual
 
@@ -41,7 +41,7 @@ def parse_raw_output_neb(out_file, input_dict,parser_opts=None):
 
     job_successful = True
     parser_info = get_parser_info(parser_info_template='aiida-quantumespresso parser neb.x v{}')
-    
+
     # load NEB out file
     try:
         with open(out_file,'r') as f:
@@ -60,11 +60,11 @@ def parse_raw_output_neb(out_file, input_dict,parser_opts=None):
             break
     if not finished_run: # error if the job has not finished
         warning = 'QE neb run did not reach the end of the execution.'
-        parser_info['parser_warnings'].append(warning)        
+        parser_info['parser_warnings'].append(warning)
         job_successful = False
 
     # parse the text output of the neb calculation
-    try: 
+    try:
         out_data,iteration_data,critical_messages = parse_neb_text_output(out_lines,input_dict)
     except QEOutputParsingError:
         if not finished_run: # I try to parse it as much as possible
@@ -81,9 +81,9 @@ def parse_raw_output_neb(out_file, input_dict,parser_opts=None):
     tmp_iteration_data = copy.copy(iteration_data)
     for x in six.iteritems(tmp_iteration_data):
         if x[0] in skip_keys:
-            continue    
+            continue
         out_data[x[0]] = x[1][-1]
-            
+
     # if there is a severe error, the calculation is FAILED
     if any([ x in out_data['warnings'] for x in critical_messages]):
         job_successful = False
@@ -98,39 +98,39 @@ def parse_raw_output_neb(out_file, input_dict,parser_opts=None):
 def parse_neb_text_output(data,input_dict={}):
     """
     Parses the text output of QE Neb.
-    
+
     :param data: a string, the file as read by read()
     :param input_dict: dictionary with the input parameters
-    
-    :return parsed_data: dictionary with key values, referring to quantities 
+
+    :return parsed_data: dictionary with key values, referring to quantities
                          at the last step.
     :return iteration_data: key,values referring to intermediate iterations.
                              Empty dictionary if no value is present.
     :return critical_messages: a list with critical messages. If any is found in
                                parsed_data['warnings'], the calculation is FAILED!
     """
-    from aiida_quantumespresso.parsers.raw_parser_pw import parse_QE_errors
+    from aiida_quantumespresso.parsers.parse_raw.pw import parse_QE_errors
     from collections import defaultdict
-    
+
     # TODO: find a more exhaustive list of the common errors of neb
-    
+
     # critical warnings: if any is found, the calculation status is FAILED
     critical_warnings = {'scf convergence NOT achieved on image':
                          'SCF did not converge for a given image',
                          'Maximum CPU time exceeded':'Maximum CPU time exceeded',
-                         'reached the maximum number of steps': 'Maximum number of iterations reached in the image optimization',       
+                         'reached the maximum number of steps': 'Maximum number of iterations reached in the image optimization',
                          '%%%%%%%%%%%%%%':None,
                          }
-    
-    minor_warnings = {'Warning:':None,                           
+
+    minor_warnings = {'Warning:':None,
                       }
-    
+
     all_warnings = dict(list(critical_warnings.items()) + list(minor_warnings.items()))
-    
+
     parsed_data = {}
     parsed_data['warnings'] = []
     iteration_data = defaultdict(list)
-    
+
     # parse time, starting from the end
     # apparently, the time is written multiple times
     for line in reversed(data.split('\n')):
@@ -140,7 +140,7 @@ def parse_neb_text_output(data,input_dict={}):
                 parsed_data['wall_time'] = time
             except Exception:
                 parsed_data['warnings'].append('Error while parsing wall time.')
-                
+
             try:
                 parsed_data['wall_time_seconds'] = \
                     convert_qe_time_to_sec(parsed_data['wall_time'])
@@ -148,9 +148,9 @@ def parse_neb_text_output(data,input_dict={}):
                 raise QEOutputParsingError("Unable to convert wall_time in seconds.")
             break
 
-    # set by default the calculation as not converged.     
+    # set by default the calculation as not converged.
     parsed_data['converged'] = [False,0]
-    
+
     for count, line in enumerate(data.split('\n')):
         if 'initial path length' in line:
             initial_path_length = float(line.split('=')[1].split('bohr')[0])
@@ -193,12 +193,12 @@ def parse_neb_text_output(data,input_dict={}):
         elif any( i in line for i in all_warnings):
             message = [ all_warnings[i] for i in all_warnings.keys() if i in line][0]
             if message is None:
-                message = line                
-                    
+                message = line
+
             if '%%%%%%%%%%%%%%' in line:
                 message  = None
-                messages = parse_QE_errors(data.split('\n'),count,parsed_data['warnings']) 
-                        
+                messages = parse_QE_errors(data.split('\n'),count,parsed_data['warnings'])
+
             # if it found something, add to log
             try:
                 parsed_data['warnings'].extend(messages)
@@ -218,30 +218,30 @@ def parse_neb_text_output(data,input_dict={}):
 
     iteration_lines = data.split('-- iteration')[1:]
     iteration_lines = [ i.split('\n') for i in iteration_lines]
-    
+
     for iteration in iteration_lines:
         for count, line in enumerate(iteration):
-            if 'activation energy (->)' in line:            
-                activ_energy = float(line.split('=')[1].split('eV')[0])                
-                iteration_data['forward_activation_energy'].append(activ_energy)
-            elif 'activation energy (<-)' in line:            
+            if 'activation energy (->)' in line:
                 activ_energy = float(line.split('=')[1].split('eV')[0])
-                iteration_data['backward_activation_energy'].append(activ_energy)   
+                iteration_data['forward_activation_energy'].append(activ_energy)
+            elif 'activation energy (<-)' in line:
+                activ_energy = float(line.split('=')[1].split('eV')[0])
+                iteration_data['backward_activation_energy'].append(activ_energy)
             elif 'image        energy (eV)        error (eV/A)        frozen' in line:
                 energies = []
                 forces = []
-                frozen = []        
+                frozen = []
                 try:
                     for i in range(num_images):
                         split_line = iteration[count+2+i].split()[1:]
                         energies.append(float(split_line[0]))
                         forces.append(float(split_line[1]))
-                        frozen.append(True if split_line[2] == 'T' else False)                        
+                        frozen.append(True if split_line[2] == 'T' else False)
                     iteration_data['image_energies'].append(energies)
                     iteration_data['image_forces'].append(forces)
-                    iteration_data['image_frozen'].append(frozen) 
+                    iteration_data['image_frozen'].append(frozen)
                 except Exception:
-                        parsed_data['warnings'].append('Error while parsing the image energies and forces.')  
+                        parsed_data['warnings'].append('Error while parsing the image energies and forces.')
             elif 'climbing image' in line:
                 iteration_data['climbing_image_auto'].append([ int(_) for _  in line.split('=')[1].split(',')])
             elif 'path length' in line:
@@ -250,5 +250,5 @@ def parse_neb_text_output(data,input_dict={}):
             elif 'inter-image distance' in line:
                 image_dist = float(line.split('=')[1].split('bohr')[0])
                 iteration_data['image_dist'].append(image_dist * bohr_to_ang)
-                
+
     return parsed_data, dict(iteration_data), list(critical_warnings.values())
