+++ docs/source/module_guide/parsers.rst
@@ -21,15 +21,7 @@ Pw Parser
 Raw Pw Parser
 -------------
 
-.. automodule:: aiida_quantumespresso.parsers.raw_parser_pw
-   :members:
-
-Basic Raw Pw Parser
--------------------
-
-This is just a minimal parser for basic quantities.
-
-.. automodule:: aiida_quantumespresso.parsers.basic_raw_parser_pw
+.. automodule:: aiida_quantumespresso.parsers.parse_raw.pw
    :members:
 
 Cp Parser
@@ -41,17 +33,8 @@ Cp Parser
 Raw Cp Parser
 -------------
 
-.. automodule:: aiida_quantumespresso.parsers.raw_parser_cp
+.. automodule:: aiida_quantumespresso.parsers.parse_raw.cp
    :members:
-
-Basic Raw Cp Parser
--------------------
-
-This is just a minimal parser for basic quantities.
-
-.. automodule:: aiida_quantumespresso.parsers.basic_raw_parser_cp
-   :members:
-
 
 Constants
 ---------
